+++ relm4-macros/src/widgets/mod.rs
@@ -1,13 +1,34 @@
-use proc_macro2::Span as Span2;
 use proc_macro2::TokenStream as TokenStream2;
-use syn::{
-    punctuated::Punctuated, spanned::Spanned, token, Expr, ExprClosure, Generics, Ident, Lit, Path,
-};
+use syn::{punctuated::Punctuated, token, Expr, ExprClosure, Generics, Ident, Lit, Path};
 
 use crate::args::Args;
 
 mod gen;
 mod parse;
+
+#[derive(Debug, Default)]
+pub struct TokenStreams {
+    /// The tokens for the struct fields -> name: Type,
+    pub struct_fields: TokenStream2,
+    /// The tokens initializing the widgets.
+    pub init_widgets: TokenStream2,
+    /// The tokens connecting widgets.
+    pub connect_widgets: TokenStream2,
+    /// The tokens initializing the properties.
+    pub init_properties: TokenStream2,
+    /// The tokens for the returned struct fields -> name,
+    pub return_fields: TokenStream2,
+    /// The view tokens (watch! macro)
+    pub view: TokenStream2,
+    /// The view tokens (track! macro)
+    pub track: TokenStream2,
+    /// The tokens for connecting events.
+    pub connect: TokenStream2,
+    /// The tokens for connecting events to components.
+    pub connect_components: TokenStream2,
+    /// The tokens for using the parent stream.
+    pub parent: TokenStream2,
+}
 
 #[derive(Debug)]
 pub(super) struct Tracker {
@@ -20,7 +41,7 @@ pub(super) enum PropertyType {
     Expr(Expr),
     Value(Lit),
     Track(Tracker),
-    Component(Expr),
+    Parent(Expr),
     Args(Args<Expr>),
     Connect(ExprClosure),
     ConnectComponent(ExprClosure),
@@ -35,22 +56,17 @@ pub enum PropertyName {
     Path(Path),
 }
 
-impl Spanned for PropertyName {
-    fn span(&self) -> Span2 {
-        match self {
-            PropertyName::Ident(ident) => ident.span(),
-            PropertyName::Path(path) => path.span(),
-        }
-    }
-}
-
 #[derive(Debug)]
 pub(super) struct Property {
+    /// Either a path or just an ident
     pub name: PropertyName,
     pub ty: PropertyType,
     pub generics: Option<Generics>,
+    /// Optional arguments like param_name(arg1, arg2, ...)
     pub args: Option<Args<Expr>>,
+    /// Assign with an ?
     pub optional_assign: bool,
+    /// Iterate through elements to generate tokens
     pub iterative: bool,
 }
 
@@ -66,15 +82,6 @@ pub(super) struct WidgetFunc {
     pub ty: Option<Vec<Ident>>,
 }
 
-impl Spanned for WidgetFunc {
-    fn span(&self) -> Span2 {
-        self.path_segments
-            .first()
-            .expect("Expected path segments in WidgetFunc")
-            .span()
-    }
-}
-
 #[derive(Debug)]
 pub(super) struct Widget {
     pub name: Ident,
@@ -82,17 +89,13 @@ pub(super) struct Widget {
     pub properties: Properties,
     pub wrapper: Option<Ident>,
     pub assign_as_ref: bool,
+    pub returned_widget: Option<ReturnedWidget>,
 }
 
-impl<'a> Widget {
-    pub fn get_widget_list(&'a self, widgets: &mut Vec<&'a Widget>) {
-        widgets.push(self);
-
-        for prop in &self.properties.properties {
-            let ty = &prop.ty;
-            if let PropertyType::Widget(widget) = ty {
-                widget.get_widget_list(widgets);
-            }
-        }
-    }
+#[derive(Debug)]
+pub(super) struct ReturnedWidget {
+    pub name: Ident,
+    pub ty: Option<Path>,
+    pub properties: Properties,
+    pub is_optional: bool,
 }