+++ src/lib.rs
@@ -18,6 +18,7 @@ mod app;
 mod component;
 pub mod drawing;
 pub mod factory;
+mod micro_component;
 mod msg_handler;
 mod traits;
 pub mod util;
@@ -25,6 +26,7 @@ mod worker;
 
 pub use app::RelmApp;
 pub use component::RelmComponent;
+pub use micro_component::{MicroComponent, MicroComponentError, MicroModel, MicroWidgets};
 pub use msg_handler::RelmMsgHandler;
 pub use traits::*;
 pub use util::widget_plus::WidgetPlus;